// src/facebook/types/main.ts
// Generated by https://quicktype.io

export interface FBLongLive {
  access_token: string;
  token_type: string;
  expires_in: number;
}

// Generated by https://quicktype.io

export interface FBAccessToken {
  data: Datum[];
  paging: Paging;
}

export interface Datum {
  access_token: string;
  category: string;
  category_list: CategoryList[];
  name: string;
  id: string;
  tasks: string[];
}

export interface CategoryList {
  id: string;
  name: string;
}

export interface Paging {
  cursors: Cursors;
}

export interface Cursors {
  before: string;
  after: string;
}